Added ability to transcode Facebook (#13)
Facebook Live doesn't want to receive live streams over 720P @ 4500 Kbps, 30 FPS. If you are pushing 1080P for YouTube Live and Twitch, the rebroadcast to Facebook Live needs to be transcoded first. This PR adds the optional ability to transcode the stream to 720P before pushing to Facebook. This transcoding is optional and should only be used if you are pushing above 720P.

Also fixed the transcoding configuration for Periscope, and migrated the Docker image towards using `jrottenberg/ffmpeg:4.2-alpine` as the base in order to get a version of `ffmpeg` that enables the `libfdk_aac` audio codec.