Configure RTMP rebroadcasting via a JSON file (#15)
This change fundamentally changes how the rebroadcasting of an RTMP stream is configured. Rather than setting an environment variable for every rebroadcast destination, this change now has a single JSON file in which each destination is configured. The additional benefit of doing this we can cleanly support rebroadcasting to multiple stream keys on the same platform, and enable destination specific transcoding. Documentation has been updated for these changes, and added a CHANGELOG file.

Closes #10